Add files via upload
This is done for Login RegistrationForm. On the home page, we have two buttons for signup and login. If a new user came, then we have to register with some credentials. After that, it will navigate to the login page. If authentication is done on the login page, It will show the dashboard( which includes all user's details).  All the fields have validations like invalid phone numbers, phone numbers that already exist, name invalid, and more(for both Login & Signup). 


Note: 
          1. LoginAuthenticationAngular.zip is the java zip file. We have to run on Eclipse.
          2. LoginAuthenticationByDB.zip is an angular file. We have to run on VS Code.